pyproject.toml
[project]
name = "who-knew-it"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit (>=1.41.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "imdbpy (>=2022.7.9,<2023.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
